fix:商品検索画面&ユーザ機能修正

+++ database/migrations/2020_08_21_155423_create_producs_table.php
@@ -22,7 +22,7 @@ class CreateProducsTable extends Migration
             $table->integer('sale_status_id')->unsigned()->index();
             $table->integer('product_status_id')->unsigned()->index();
             $table->timestamp('regist_date');
-            $table->integer('user_id')->unsigned()->index();
+            $table->bigInteger('user_id')->unsigned()->index();
             $table->char('delete_flag', 1);
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
